Move buffer keyword from p2p to crypto-validation
