+++ packages/p2p/source/validation/index.ts
@@ -1,3 +1,2 @@
 export * from "./formats.js";
 export * from "./is-valid-peer.js";
-export * from "./keywords.js";
