+++ packages/p2p/test/helpers/prepare-validator-context.ts
@@ -4,11 +4,11 @@ import { Configuration } from "@mainsail/crypto-config/distribution/index.js";
 import { makeKeywords as makeProposalKeywords } from "@mainsail/crypto-proposal/distribution/keywords.js";
 import { schemas as cryptoTransactionSchemas } from "@mainsail/crypto-transaction/distribution/index.js";
 import { schemas as cryptoValidationSchemas } from "@mainsail/crypto-validation/distribution/index.js";
+import { makeKeywords as makeCryptoValidationKeywords } from "@mainsail/crypto-validation/distribution/keywords.js";
 import type { Application } from "@mainsail/kernel";
 import type { Validator } from "@mainsail/validation";
 
 import cryptoJson from "../../../core/bin/config/devnet/core/crypto.json" with { type: "json" };
-import { makeKeywords } from "../../source/validation/keywords.js";
 
 type Context = {
 	app: Application;
@@ -20,10 +20,12 @@ export const prepareValidatorContext = (context: Context) => {
 	context.app.get<Configuration>(Identifiers.Cryptography.Configuration).setConfig(cryptoJson);
 	context.app.get<Configuration>(Identifiers.Cryptography.Configuration).setHeight(1);
 
-	const keywords = makeKeywords();
-	context.validator.addKeyword(keywords.buffer);
-
 	const configuration = context.app.get<Configuration>(Identifiers.Cryptography.Configuration);
+
+	for(const keyword of Object.values(makeCryptoValidationKeywords(configuration))) {
+		context.validator.addKeyword(keyword);
+	}
+
 	const messageKeywords = makeProposalKeywords(configuration);
 	context.validator.addKeyword(messageKeywords.limitToRoundValidators);
 	context.validator.addKeyword(messageKeywords.isValidatorIndex);
